Switch back to memory tables (but still load data from CSV files via text tables)
The previous version had a serious performance problem. The
database took a very long time to load because we were creating
indexes on text tables in a compressed file. So in this revision,
we go back to using memory tables (with indexes) and the text
tables are merely the source of data.

The text tables are in a new schema called "foodmart_csv". We
don't suggest using these tables for anything beside sequential
scan. The memory tables are in the "foodmart" schema as
previously.